+++ 09.JSONProcessing/ProductsShop/src/main/java/productsshop/repository/CategoryRepository.java
@@ -3,10 +3,13 @@ package productsshop.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
+import productsshop.domain.dtos.CategoryNameProductsCountAverageAndTotalPricesDto;
 import productsshop.domain.entities.Category;
 
+import java.util.List;
+
 @Repository
-public interface CategoryRepository extends JpaRepository<Category, Long> {
-    
+public interface CategoryRepository extends JpaRepository<Category, Integer> {
+
 
 }
